Moved and fixed tests
